+++ module/hastebin/src/main/kotlin/sibyl/haste/HasteServerResponse.kt
@@ -1,0 +1,8 @@
+package sibyl.haste
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class HasteServerResponse(
+    val key: String
+)